# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: Noisia Public Reporting API
  version: "2.0.0"
  summary: Public reporting endpoints for customers and external integrations.
  description: |
    Public API for published Noisia Signal reports.

    V1 exposes flat tabular datasets and CSV exports for BI tools.
    V2 exposes the structured client-ready Signal output for external integrations.
servers:
  - url: https://studio.noisia.ai
    description: Production
security:
  - bearerAuth: []
  - noisiaApiKey: []
...
paths:
...
  /api/data-os/signal/{workspaceId}/triggers-barriers:
    get:
      tags: [Signal Workspace V1]
      operationId: getSignalWorkspaceTriggersBarriersOverview
      summary: Read the current T&B report revision by workspace and report key.
      description: Reads relational findings, artifacts and snapshot identity. It never serves published_outputs.payload and does not require a study query parameter.
      security: [{ kindeSession: [] }]
      parameters:
        - { $ref: "#/components/parameters/WorkspaceId" }
        - { $ref: "#/components/parameters/SignalStart" }
        - { $ref: "#/components/parameters/SignalEnd" }
      responses:
        "200": { $ref: "#/components/responses/SignalWorkspaceSuccess" }
        "304": { description: ETag still current. }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/SignalBackendError" }
        "503": { $ref: "#/components/responses/SignalBackendError" }
...
  /api/data-os/signal/{workspaceId}/triggers-barriers/findings/{findingId}:
    get:
      tags: [Signal Workspace V1]
      operationId: getSignalWorkspaceTriggersBarriersFinding
      summary: Read one finding from the current report revision.
      security: [{ kindeSession: [] }]
      parameters:
        - { $ref: "#/components/parameters/WorkspaceId" }
        - name: findingId
          in: path
          required: true
          schema: { type: string, format: uuid }
        - { $ref: "#/components/parameters/SignalStart" }
        - { $ref: "#/components/parameters/SignalEnd" }
      responses:
        "200": { $ref: "#/components/responses/SignalWorkspaceSuccess" }
        "304": { description: ETag still current. }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/SignalBackendError" }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: API key
      description: "Preferred auth header: `Authorization: Bearer <NOISIA_REPORTING_API_KEY>`."
    noisiaApiKey:
      type: apiKey
      in: header
      name: x-noisia-api-key
      description: Fallback API key header.
    kindeSession:
      type: apiKey
      in: cookie
      name: noisia_studio_session
      description: Logical Studio session; Kinde authenticates and Noisia DB authorizes workspace and brand access.
  parameters:
...
    WorkspaceId:
      name: workspaceId
      in: path
      required: true
      schema: { type: string, format: uuid }
      description: Stable Signal workspace id, never an output id.
...
    SignalStart:
      name: start
      in: query
      required: true
      schema: { type: string, format: date }
    SignalEnd:
      name: end
      in: query
      required: true
      schema: { type: string, format: date }
...
  responses:
    SignalWorkspaceSuccess:
      description: signal-backend-v1 response. Concrete shape depends on the route.
      content:
        application/json:
          schema:
            type: object
            required: [contract_version]
            properties:
              contract_version: { const: signal-backend-v1 }
            additionalProperties: true
...
    SignalBackendError:
      description: Typed signal-backend-v1 contract error.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/SignalBackendErrorV1" }
    Unauthorized:
      description: Missing or invalid API key.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Forbidden:
      description: API key cannot access this report.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
  schemas:
...
    SignalBackendErrorV1:
      type: object
      required: [contract_version, error, message, details]
      properties:
        contract_version: { const: signal-backend-v1 }
        error: { type: string, enum: [invalid_filter, unsupported_dimension, stale, partial, not_available] }
        message: { type: string }
        details: { type: object, additionalProperties: true }
...
    ErrorResponse:
      type: object
      required: [error, message]
      properties:
        error:
          type: string
          examples: [missing_api_key, invalid_api_key, forbidden_output, report_not_found, server_error]
        message:
          type: string
...